+++ src/Status.php
@@ -153,7 +153,7 @@ class Status extends RobotsBase
     private function processRedirect()
     {
         if (!isset($this->originalUrl)) {
-            $this->originalUrl = $this->url;
+            $this->setOriginalUrl($this->url);
         }
 
         if (count($this->redirects) > 10) {
@@ -164,5 +164,30 @@ class Status extends RobotsBase
         $this->redirects[] = "{$this->statusCode} {$redirectUrl}";
         $this->setURL($redirectUrl)->setNotFetched();
         return $this->validate();
+    }
+
+    public function getRedirects($includeOriginal = false)
+    {
+        $redirects = $this->redirects;
+        if ($includeOriginal) {
+            array_unshift($redirects, "Original URL: {$this->originalUrl}");
+        }
+
+        return $redirects;
+    }
+
+    public function setRedirects($redirects)
+    {
+        $this->redirects = $redirects;
+    }
+
+    public function getOriginalUrl()
+    {
+        return $this->originalUrl;
+    }
+
+    public function setOriginalUrl($url)
+    {
+        $this->originalUrl = $url;
     }
 }